+++ src/main/java/org/apache/aurora/scheduler/base/TaskTestUtil.java
@@ -62,7 +62,7 @@ import org.apache.mesos.v1.Protos.ExecutorInfo;
  */
 public final class TaskTestUtil {
 
-  public static final IJobKey JOB = JobKeys.from("role", "dev", "job");
+  public static final IJobKey JOB = JobKeys.from("role", "devel", "job");
   public static final TierInfo REVOCABLE_TIER =
       new TierInfo(true /* preemptible */, true /* revocable */);
   public static final TierInfo DEV_TIER =
@@ -86,7 +86,8 @@ public final class TaskTestUtil {
           true,
           true,
           true,
-          true);
+          true,
+          ConfigurationManager.DEFAULT_ALLOWED_JOB_ENVIRONMENTS);
   public static final ExecutorID EXECUTOR_ID = ExecutorID.newBuilder()
       .setValue("PLACEHOLDER")
       .build();