+++ src/test/java/org/apache/aurora/scheduler/configuration/ConfigurationManagerTest.java
@@ -44,6 +44,7 @@ import org.apache.aurora.scheduler.configuration.ConfigurationManager.Configurat
 import org.apache.aurora.scheduler.configuration.ConfigurationManager.TaskDescriptionException;
 import org.apache.aurora.scheduler.mesos.TestExecutorSettings;
 import org.apache.aurora.scheduler.storage.entities.IDockerParameter;
+import org.apache.aurora.scheduler.storage.entities.IJobConfiguration;
 import org.apache.aurora.scheduler.storage.entities.ITaskConfig;
 import org.junit.Rule;
 import org.junit.Test;
@@ -124,7 +125,8 @@ public class ConfigurationManagerTest {
           true,
           false,
           true,
-          false),
+          false,
+          ConfigurationManager.DEFAULT_ALLOWED_JOB_ENVIRONMENTS),
       TaskTestUtil.TIER_MANAGER,
       TaskTestUtil.THRIFT_BACKFILL,
       TestExecutorSettings.THERMOS_EXECUTOR);
@@ -136,7 +138,8 @@ public class ConfigurationManagerTest {
           false,
           true,
           true,
-          true),
+          true,
+          ConfigurationManager.DEFAULT_ALLOWED_JOB_ENVIRONMENTS),
       TaskTestUtil.TIER_MANAGER,
       TaskTestUtil.THRIFT_BACKFILL,
       TestExecutorSettings.THERMOS_EXECUTOR);
@@ -289,7 +292,8 @@ public class ConfigurationManagerTest {
             false,
             false,
             false,
-            false),
+            false,
+            ConfigurationManager.DEFAULT_ALLOWED_JOB_ENVIRONMENTS),
         TaskTestUtil.TIER_MANAGER,
         TaskTestUtil.THRIFT_BACKFILL,
         TestExecutorSettings.THERMOS_EXECUTOR).validateAndPopulate(ITaskConfig.build(builder));
@@ -312,7 +316,8 @@ public class ConfigurationManagerTest {
                     false,
                     false,
                     false,
-                    false),
+                    false,
+                    ".+"),
             TaskTestUtil.TIER_MANAGER,
             TaskTestUtil.THRIFT_BACKFILL,
             TestExecutorSettings.THERMOS_EXECUTOR).validateAndPopulate(ITaskConfig.build(builder));
@@ -340,6 +345,27 @@ public class ConfigurationManagerTest {
     ITaskConfig populated =
         DOCKER_CONFIGURATION_MANAGER.validateAndPopulate(ITaskConfig.build(builder));
     assertEquals(ImmutableSet.of("health", "http"), populated.getTaskLinks().keySet());
+  }
+
+  @Test
+  public void testJobEnvironmentValidation() throws Exception {
+    JobConfiguration jobConfiguration = UNSANITIZED_JOB_CONFIGURATION.deepCopy();
+    jobConfiguration.getKey().setEnvironment("foo");
+    expectTaskDescriptionException("Job environment foo doesn't match: b.r");
+    new ConfigurationManager(
+      new ConfigurationManagerSettings(
+          ALL_CONTAINER_TYPES,
+          true,
+          ImmutableList.of(new DockerParameter("foo", "bar")),
+          false,
+          true,
+          true,
+          true,
+          "b.r"),
+      TaskTestUtil.TIER_MANAGER,
+      TaskTestUtil.THRIFT_BACKFILL,
+      TestExecutorSettings.THERMOS_EXECUTOR)
+            .validateAndPopulate(IJobConfiguration.build(jobConfiguration));
   }
 
   private void expectTaskDescriptionException(String message) {