+++ src/test/java/org/apache/aurora/scheduler/cron/quartz/CronIT.java
@@ -59,7 +59,7 @@ import static org.junit.Assert.assertTrue;
 public class CronIT extends EasyMockTest {
   public static final CrontabEntry CRONTAB_ENTRY = CrontabEntry.parse("* * * * *");
 
-  private static final IJobKey JOB_KEY = JobKeys.from("roll", "b", "c");
+  private static final IJobKey JOB_KEY = JobKeys.from("roll", "prod", "c");
   private static final Identity IDENTITY = new Identity().setUser("user");
 
   private static final IJobConfiguration CRON_JOB = IJobConfiguration.build(